♻️ Huge refactoring
