+++ banana/__init__.py
@@ -1,4 +1,4 @@
-from .main import Banana, refresh
+from .main import Banana
 
 
 __version__ = "0.5.0"
