+++ banana/core/history.py
@@ -18,7 +18,7 @@ from sqlalchemy import (
 from sqlalchemy.ext.declarative import declarative_base
 from sqlalchemy.orm import sessionmaker
 
-from .instances import config
+from .config import config
 
 
 Base = declarative_base()
