+++ banana/layout.py
@@ -3,7 +3,7 @@ from dash_ag_grid import AgGrid
 from dash_iconify import DashIconify
 import dash_mantine_components as dmc
 
-from .core.instances import config
+from .core.config import config
 
 
 class Layout(dmc.MantineProvider):
@@ -26,18 +26,7 @@ class Layout(dmc.MantineProvider):
 
     def left_section(self) -> html.Div:
         return dmc.Paper(
-            children=[
-                html.Div(id="banana--menu"),
-                dmc.Button(
-                    "Refresh",
-                    id="banana--refresh-button",
-                    color=config.theme,
-                    radius="md",
-                    leftSection=DashIconify(icon="mingcute:refresh-3-fill", height=20),
-                    variant="light",
-                    mt=20,
-                ),
-            ],
+            children=[html.Div(id="banana--menu")],
             className="left-section",
             bg=dmc.DEFAULT_THEME["colors"][config.theme][9],
         )
