+++ banana/main.py
@@ -22,23 +22,15 @@ from .callbacks import (
     UpdateCellCallback,
 )
 from .layout import Layout
-from .core.compile import Compiler
-from .core.instances import config, server
+from .core.config import config, server
 from .core.utils import raise_error
 
 
 _dash_renderer._set_react_version("18.2.0")
 
 
-def refresh():
-    with server.app_context():
-        obj = Compiler()
-        obj.refresh()
-
-
 class Banana(Dash):
     def __init__(self):
-        refresh()
         super().__init__(
             server=server,
             assets_folder=resources.files("banana") / "assets",
@@ -97,16 +89,10 @@ class Banana(Dash):
 
         @self.callback(
             Output("banana--menu", "children"),
-            Input("banana--location", "pathname"),
-            Input("banana--refresh-button", "n_clicks"),
+            Input("banana--menu", "style"),
+            State("banana--location", "pathname"),
         )
-        def load_side_menu(pathname: str, _):
-            if ctx.triggered_id == "banana--refresh-button":
-                try:
-                    refresh()
-                except Exception as e:
-                    raise_error("Error on refreshing table configuration", str(e))
-
+        def load_side_menu(_, pathname: str):
             obj = LoadSideMenuCallback(pathname)
             return obj.menu
