+++ banana/queries/insert_row.py
@@ -2,8 +2,8 @@ from sqlalchemy import Table, MetaData
 from sqlalchemy.sql import insert
 
 from .execute_query import exec_sql
-from ..core.instances import db
-from ..models.table_models import BananaTable
+from ..core.config import db
+from ..core.tables import BananaTable
 
 
 def insert_row(banana_table: BananaTable, values: dict) -> None:
